requirements.txt
gdown

wordcloud
pandas
numpy
matplotlib
seaborn
collections

coloredlogs
transformers
pytorch
kaggle
